refactored weapon and shooting
- created a weapon manager component
- refactored shooting logic to include muzzle flash
- refactored shooting properties resources
