+++ Components/Scripts/ShootingComponent.cs
@@ -9,17 +9,15 @@ namespace ElGatoProject.Components.Scripts;
 [GlobalClass]
 public partial class ShootingComponent : Node2D
 {
-	[Export] public Utility.PlayerOrEnemy PlayerOrEnemy;
-	[Export] public Utility.WeaponType WeaponType;
+	// [Export] public Utility.PlayerOrEnemy PlayerOrEnemy;
+	// [Export] public Utility.WeaponType WeaponType;
 	[Export] public ShootingProperties ShootingProperties;
 	[Export] private Marker2D _muzzle;
+	[Export] private AnimatedSprite2D _flashSprite;
 	[Export] private Timer _shotCooldownTimer, _reloadTimer;
 
 	[Signal]
 	public delegate void ShootingEventHandler();
-
-	[Signal]
-	public delegate void CannotShootEventHandler();
 
 	public bool HurtStatus;
 	private bool _onCooldown, _reloading;
@@ -30,32 +28,34 @@ public partial class ShootingComponent : Node2D
 	{
 		SetTimerValues();
 		ConnectToSignals();
-		
+
 		if (_muzzle != null) _muzzlePosition = _muzzle.Position;
+		
+		if (ShootingProperties != null) SetWeaponMuzzleFlash(ShootingProperties.WeaponType);
+		
 	}
 	
-	// Public method
+	// Public methods
 	public bool Shoot(Vector2 targetVector)
 	{
-		bool shootStatus;
+		var shootStatus = false;
 		if (!_onCooldown)
 		{
 			EmitSignal(SignalName.Shooting);
+			
+			_flashSprite.Play(Utility.EntityAnimations[Utility.EntityState.Shoot]);
+			
 			ShootingLogic(targetVector);
+			
 			_onCooldown = true;
 			_shotCooldownTimer?.Start();
+			
 			shootStatus = true;
-		}
-		else
-		{
-			EmitSignal(SignalName.CannotShoot);
-			shootStatus = false;
 		}
 
 		return shootStatus;
 	}
-
-	// Helper methods
+	
 	public void SetTimerValues()
 	{
 		if (_shotCooldownTimer != null)
@@ -70,7 +70,22 @@ public partial class ShootingComponent : Node2D
 			if (ShootingProperties != null) _reloadTimer.WaitTime = ShootingProperties.ReloadTime;
 		}
 	}
+	
+	public void SetWeaponMuzzleFlash(Utility.WeaponType weaponType)
+	{
+		// Remove current muzzle flash sprite
+		if (_flashSprite != null)
+			_muzzle?.RemoveChild(_flashSprite); 
+				
+		// Add new muzzle flash sprite
+		_flashSprite = ResourceLoader
+			.Load<PackedScene>(Utility.MuzzleFlashPackedScenePaths[weaponType])
+			.Instantiate<AnimatedSprite2D>();
+		
+		_muzzle?.AddChild(_flashSprite);
+	}
 
+	// Signals
 	private void ConnectToSignals()
 	{
 		if (_shotCooldownTimer != null) _shotCooldownTimer.Timeout += OnShotCoolDownTimerTimeout;
@@ -88,29 +103,13 @@ public partial class ShootingComponent : Node2D
 		_reloading = false;
 		_bulletCount = 0;
 	}
-
-	private void FlipMuzzle(Vector2 targetVector)
-	{
-		switch (targetVector.X)
-		{
-			case < 0:
-			{
-				if (_muzzle != null) _muzzle.Position = new Vector2(-_muzzlePosition.X, _muzzlePosition.Y);
-				break;
-			}
-			case > 0:
-			{
-				if (_muzzle != null) _muzzle.Position = new Vector2(_muzzlePosition.X, _muzzlePosition.Y);
-				break;
-			}
-		}
-	}
-
+	
+	// Helper methods
 	private void ShootingLogic(Vector2 targetVector)
 	{
 		FlipMuzzle(targetVector);
 		
-		switch (WeaponType)
+		switch (ShootingProperties.WeaponType)
 		{
 			case Utility.WeaponType.None:
 				break;
@@ -118,7 +117,8 @@ public partial class ShootingComponent : Node2D
 			case Utility.WeaponType.EnemyShotgun:
 				for (int i = 0; i < ShootingProperties?.BulletsPerShot; i++)
 				{
-					CreateAndSetBulletProperties(WeaponType, GlobalPosition.DirectionTo(targetVector), i);
+					CreateAndSetBulletProperties(ShootingProperties.WeaponType,
+						GlobalPosition.DirectionTo(targetVector), i);
 				}
 				break;
 			case Utility.WeaponType.EnemyPistol:
@@ -126,7 +126,8 @@ public partial class ShootingComponent : Node2D
 			case Utility.WeaponType.EnemyRailGun:
 				if (!_reloading)
 				{
-					CreateAndSetBulletProperties(WeaponType, GlobalPosition.DirectionTo(targetVector));
+					CreateAndSetBulletProperties(ShootingProperties.WeaponType,
+						GlobalPosition.DirectionTo(targetVector));
 					
 					_bulletCount++;
 					
@@ -141,15 +142,36 @@ public partial class ShootingComponent : Node2D
 			case Utility.WeaponType.PlayerShotgun:
 				for (int i = 0; i < ShootingProperties?.BulletsPerShot; i++)
 				{
-					CreateAndSetBulletProperties(WeaponType, new Vector2(targetVector.X, targetVector.Y), i);
+					CreateAndSetBulletProperties(ShootingProperties.WeaponType,
+						new Vector2(targetVector.X, targetVector.Y), i);
 				}
 				break;
 			
 			case Utility.WeaponType.PlayerPistol:
 			case Utility.WeaponType.PlayerMachineGun:
 			case Utility.WeaponType.PlayerRailGun:
-				CreateAndSetBulletProperties(WeaponType, new Vector2(targetVector.X, targetVector.Y));
+				CreateAndSetBulletProperties(ShootingProperties.WeaponType,
+					new Vector2(targetVector.X, targetVector.Y));
 				break;
+		}
+	}
+	
+	private void FlipMuzzle(Vector2 targetVector)
+	{
+		switch (targetVector.X)
+		{
+			case < 0:
+			{
+				if (_muzzle != null) _muzzle.Position = new Vector2(-_muzzlePosition.X, _muzzlePosition.Y);
+				if (_flashSprite != null) _flashSprite.FlipH = true;
+				break;
+			}
+			case > 0:
+			{
+				if (_muzzle != null) _muzzle.Position = new Vector2(_muzzlePosition.X, _muzzlePosition.Y);
+				if (_flashSprite != null) _flashSprite.FlipH = false;
+				break;
+			}
 		}
 	}
 
@@ -159,11 +181,10 @@ public partial class ShootingComponent : Node2D
 		int projectileIndex = 0
 		)
 	{
-		// var projectileInstance = Globals.Instance.BulletProjectile.Instantiate<BulletProjectile>();
 		var projectileInstance = ResourceLoader.Load<PackedScene>(Utility.BulletPackedScenePath)
 			.Instantiate<BulletProjectile>();
 		
-		projectileInstance.PlayerOrEnemyBullet = PlayerOrEnemy;
+		projectileInstance.PlayerOrEnemyBullet = ShootingProperties.PlayerOrEnemy;
 		projectileInstance.BulletWeaponType = projectileWeaponType;
 		
 		projectileInstance.Target = directionToTarget;
@@ -198,4 +219,5 @@ public partial class ShootingComponent : Node2D
 			}
 		}
 	}
+	
 }
