+++ Players/Scenes/player_the_cat.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=116 format=3 uid="uid://c4o8x6lc2silu"]
+[gd_scene load_steps=103 format=3 uid="uid://c4o8x6lc2silu"]
 
 [ext_resource type="Script" path="res://Players/Scripts/PlayerTheCat.cs" id="1_ngcol"]
 [ext_resource type="Script" path="res://Components/Scripts/HealthComponent.cs" id="2_irjyl"]
@@ -7,11 +7,9 @@
 [ext_resource type="Script" path="res://Components/Scripts/PickupsComponent.cs" id="4_mdlkx"]
 [ext_resource type="Script" path="res://Components/Scripts/VelocityComponent.cs" id="5_jql4e"]
 [ext_resource type="Resource" uid="uid://c660ff6m6ajmy" path="res://Resources/PlayerElGatoVelocityProperties.tres" id="7_u5ntm"]
-[ext_resource type="Script" path="res://Players/Scripts/WeaponElgato.cs" id="9_ion02"]
 [ext_resource type="Texture2D" uid="uid://b6e82lhna6ymn" path="res://Assets/Players/player-001-spritesheet-003.png" id="9_mv6h0"]
-[ext_resource type="Texture2D" uid="uid://cummmt1lb6olu" path="res://Assets/Players/player-001-spritesheet-002.png" id="10_002i0"]
+[ext_resource type="Script" path="res://Components/Scripts/WeaponManagerComponent.cs" id="10_5oihl"]
 [ext_resource type="Script" path="res://Components/Scripts/ShootingComponent.cs" id="11_udspu"]
-[ext_resource type="Resource" uid="uid://cs6j3p5h3n34h" path="res://Resources/PlayerPistolShootingProperties.tres" id="13_br2kc"]
 [ext_resource type="Resource" uid="uid://dj4480r3q8euy" path="res://Resources/PlayerMachineGunShootingProperties.tres" id="13_pu5mw"]
 
 [sub_resource type="CapsuleShape2D" id="CapsuleShape2D_431m1"]
@@ -588,89 +586,6 @@ animations = [{
 "speed": 10.0
 }]
 
-[sub_resource type="AtlasTexture" id="AtlasTexture_pj2gm"]
-atlas = ExtResource("10_002i0")
-region = Rect2(0, 16, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_0eyut"]
-atlas = ExtResource("10_002i0")
-region = Rect2(32, 16, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_0oncp"]
-atlas = ExtResource("10_002i0")
-region = Rect2(64, 16, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_grswh"]
-atlas = ExtResource("10_002i0")
-region = Rect2(96, 16, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_80qu8"]
-atlas = ExtResource("10_002i0")
-region = Rect2(128, 16, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_0r7yd"]
-atlas = ExtResource("10_002i0")
-region = Rect2(0, 48, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_d87pe"]
-atlas = ExtResource("10_002i0")
-region = Rect2(32, 48, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_d1emn"]
-atlas = ExtResource("10_002i0")
-region = Rect2(64, 48, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_g1du0"]
-atlas = ExtResource("10_002i0")
-region = Rect2(96, 48, 32, 16)
-
-[sub_resource type="AtlasTexture" id="AtlasTexture_mxvsk"]
-atlas = ExtResource("10_002i0")
-region = Rect2(128, 48, 32, 16)
-
-[sub_resource type="SpriteFrames" id="SpriteFrames_ld3wl"]
-animations = [{
-"frames": [{
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_pj2gm")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_0eyut")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_0oncp")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_grswh")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_80qu8")
-}],
-"loop": true,
-"name": &"idle",
-"speed": 7.0
-}, {
-"frames": [{
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_0r7yd")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_d87pe")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_d1emn")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_g1du0")
-}, {
-"duration": 1.0,
-"texture": SubResource("AtlasTexture_mxvsk")
-}],
-"loop": false,
-"name": &"shoot",
-"speed": 15.0
-}]
-
 [sub_resource type="Animation" id="Animation_viujf"]
 length = 0.001
 tracks/0/type = "value"
@@ -1158,22 +1073,17 @@ monitoring = false
 shape = SubResource("RectangleShape2D_l4g6g")
 debug_color = Color(0.932818, 0.259544, 0, 0.42)
 
-[node name="PlayerControllerComponent" type="Node" parent="." node_paths=PackedStringArray("_playerCharacter", "_sprite", "_velocityComponent", "_leftWallDetect", "_rightWallDetect", "_dashCooldownTimer", "_dashTimer", "_gameOverTimer", "_invincibilityTimer", "_blinkTimer", "_hurtboxComponent", "_healthComponent", "_animaationPlayer", "_weapon", "_shootingComponent")]
+[node name="PlayerControllerComponent" type="Node" parent="." node_paths=PackedStringArray("_playerCharacter", "_characterSprite", "_velocityComponent", "_leftWallDetect", "_rightWallDetect", "_hurtboxComponent", "_healthComponent", "_animaationPlayer", "_weapon", "_shootingComponent")]
 script = ExtResource("2_y5cke")
 _playerCharacter = NodePath("..")
-_sprite = NodePath("../sprite")
+_characterSprite = NodePath("../sprite")
 _velocityComponent = NodePath("../VelocityComponent")
 _leftWallDetect = NodePath("../LeftWallDetect")
 _rightWallDetect = NodePath("../RightWallDetect")
-_dashCooldownTimer = NodePath("dashCooldownTimer")
-_dashTimer = NodePath("dashTimer")
-_gameOverTimer = NodePath("gameOverTimer")
-_invincibilityTimer = NodePath("invincibilityTimer")
-_blinkTimer = NodePath("blinkTimer")
 _hurtboxComponent = NodePath("../HurtboxComponent")
 _healthComponent = NodePath("../HealthComponent")
 _animaationPlayer = NodePath("../AnimationPlayer")
-_weapon = NodePath("../Weapon_ElGato")
+_weapon = NodePath("")
 _shootingComponent = NodePath("../ShootingComponent")
 
 [node name="dashCooldownTimer" type="Timer" parent="PlayerControllerComponent"]
@@ -1239,39 +1149,24 @@ debug_color = Color(0, 0.65098, 0, 0.419608)
 sprite_frames = SubResource("SpriteFrames_w0g3d")
 animation = &"idle"
 
-[node name="Weapon_ElGato" type="Node2D" parent="." node_paths=PackedStringArray("_characterSprite", "_playerController", "_pickupsComponent") groups=["Players"]]
-position = Vector2(4, 1)
-script = ExtResource("9_ion02")
-_characterSprite = NodePath("../sprite")
-_playerController = NodePath("../PlayerControllerComponent")
-_pickupsComponent = NodePath("../PickupsComponent")
-
-[node name="weaponSprite" type="AnimatedSprite2D" parent="Weapon_ElGato"]
-visible = false
-sprite_frames = SubResource("SpriteFrames_ld3wl")
-animation = &"shoot"
-
-[node name="ShootingComponent" type="Node2D" parent="Weapon_ElGato" node_paths=PackedStringArray("_muzzle", "_shotCooldownTimer")]
-script = ExtResource("11_udspu")
-WeaponType = 7
-ShootingProperties = ExtResource("13_pu5mw")
-_muzzle = NodePath("../../muzzle")
-_shotCooldownTimer = NodePath("../../ShootingComponent/shotCooldownTimer")
-
 [node name="AnimationPlayer" type="AnimationPlayer" parent="."]
 libraries = {
 "": SubResource("AnimationLibrary_ymrhi")
 }
 blend_times = [&"RESET", &"idle", 0.5, &"idle", &"idle", 2.0, &"idle_shoot", &"run_shoot", 0.1]
 
+[node name="WeaponManagerComponent" type="Node" parent="." node_paths=PackedStringArray("_pickupsComponent", "_shootingComponent")]
+script = ExtResource("10_5oihl")
+_pickupsComponent = NodePath("../PickupsComponent")
+_shootingComponent = NodePath("../ShootingComponent")
+
 [node name="ShootingComponent" type="Node2D" parent="." node_paths=PackedStringArray("_muzzle", "_shotCooldownTimer")]
 script = ExtResource("11_udspu")
-WeaponType = 5
-ShootingProperties = ExtResource("13_br2kc")
-_muzzle = NodePath("../muzzle")
+ShootingProperties = ExtResource("13_pu5mw")
+_muzzle = NodePath("muzzle")
 _shotCooldownTimer = NodePath("shotCooldownTimer")
 
 [node name="shotCooldownTimer" type="Timer" parent="ShootingComponent"]
 
-[node name="muzzle" type="Marker2D" parent="."]
+[node name="muzzle" type="Marker2D" parent="ShootingComponent"]
 position = Vector2(14, -1)
