+++ Resources/PlayerElGatoVelocityProperties.tres
@@ -7,7 +7,7 @@ script = ExtResource("1_gp16e")
 MaxSpeed = 170.0
 DashSpeed = 300.0
 Acceleration = 550.0
-Friction = 320.0
+Friction = 410.0
 JumpVelocity = -340.0
 Gravity = 780.0
 WallSlideGravity = 500.0
